Move context classes into own directory
